Rewrite scroll up.
It now does a bunch of mini scroll ups and downs.

1. Scroll up to the first physical line start above the visual start.
2. Count the number of visual lines.
3. If count > scroll amount then go to the point in the middle.
4. Otherwise decrease the scroll amount by the count and repeat.

The major benefit of this approach is that scrolling up will show the
same visual wrapping as it will once the entire line is on the screen.
Except if a tab is on the newline boundary.  I guess that's a todo.